Added all models in e-commerce database design

+++ models/src/main/java/com/android/models/result/Product.kt
@@ -39,7 +39,7 @@ data class Product(
 
     val availableColors: List<String>,
 
-    val size: Int,
+    val size: Double,
 
     val color: String,
